requirements.txt
langchain==0.0.200
python-dotenv==1.0.0
websocket_client==1.6.1
